Changed typings to alleviate build warnings, added hotkeys for scene render changes(1, 2, 3)

+++ src/D3D/StructuredBuffer.cpp
@@ -1,6 +1,6 @@
 #include "StructuredBuffer.h"
 
-StructuredBuffer::StructuredBuffer(const void* inputData, unsigned int numEle, size_t eleSize, DescriptorHeap* heap)
+StructuredBuffer::StructuredBuffer(const void* inputData, unsigned int numEle, UINT eleSize, DescriptorHeap* heap)
 	: data(inputData), descriptorHeap(heap), numElements(numEle), elementSize(eleSize)
 {
 	findFreeHandle();
